server/requirements.txt
azure-ai-textanalytics==5.1.0b3
azure-common==1.1.26
azure-core==1.9.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
oauthlib==3.1.0
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1